requirements.txt
argparse
opencv-python
tqdm
numpy
torch
segment_anything
pillow
torchvision
pycocotools
